A patch from Jeremy Olmstead for "Unable to search accounting transaction entries by productId " https://issues.apache.org/jira/browse/OFBIZ-5160
From the FindAcctgTransEntries screen, enter a part number to search for...you will still get a complete list of all transactions because the form field name is produtId instead of productId.  After searching relevant document types it turns out there are a few places where produtId is used instead of productId, the patch is meant to fix them all.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/birt@1460116 13f79535-47bb-0310-9956-ffa450edef68